Implement firefox container tab support
Firefox can have container tabs enabled, and an extension exists
that will allow routing urls with a specific format into these
containers [Open External links in a container](https://addons.mozilla.org/en-GB/firefox/addon/open-url-in-container/).

If configured in a `handler`, rewrite the outgoing URL into this format
and open that.

This has an improvement over using a rewrite rule, in that the config
can be kept DRY, and container selection logic is left to the handler.

Example configuration:

```js
handlers: [
  {
    // Some work site
    match: "https://some-site.com/*",
    browser: {
      name: "Firefox",
      container: "Work",
    },
  },
],
```